Day 5
Today was mainly the same as yesterday, and I feel that the further I get along into this hundred days of programming challenge, there will be less and less detailed notes or logs defining my experience. However, Angela said something very interesting at the end of this lecture. The idea that, when things get really difficult people start believing that they aren't meant to do something. And this has happened to me as well, programming is a ridiculously difficult skill with so many branching paths that it is overwhelming sometimes. However, with the last two months that I have spent programming every single day, and starting this course, I feel that I have gained the confidence to take on any challenge, and even if I may not be perfect, I will eventually get better.