Day 7
With the help of Angela today, I was able to create a simple hangman game. Personally, I enjoyed this project a lot and I may do some research today and try to create another game in the same vein as hangman, but either a little more complex or a little less complex. All in all a pretty good lecture for today.