Day 10
For the most part, today was also pretty good. I was able to figure out some of the more complex things however, the thing that kind of ticked me off was how I wasn't able to properly condense my code for reusability. And that's OK with more time comes more experience and things but starting tomorrow I'll try and be more vigilant about this.