Day 11
Today's assignment consisted of creating a game of blackjack. This was an exercise in working by myself, and actually figuring out how to solve a difficult problem by breaking it down into smaller and smaller chunks. For the most part, I do feel that I did a great job figuring out this assignment.