Day 35
Honestly, this main project that I had to work on today, I didn't really want to do it. I didn't really enjoy it in any way, shape or form so I did a bit of programming and tomorrow I will try to complete the other project but Angela wants me to make. However, if a project doesn't really interest me I'm gonna focus on something else to do.