Day 36, 37 and 38
Did a lot of programming. I feel that working with API's are good. However, I do need a little bit more experience being competent at it. It's not the easiest process to understand it's quite fiddly and requires a lot of in-depth knowledge. However, I do think that I am slowly getting betterAnd if I can have a little bit more experience using APIs, that would be great.