Day 41, 42 and 43
The reason why I completed three whole days of work and one is because in conjunction with 100 days of code challenge that I'm taking on, I'm also doing a web development course. This course is top by Angela so most of the content and the courses that were being taught in the past three days projects they were very easy for me to do. I already knew the answers to most of projects and the basics of HTML so yeah.