Day 45
Today I got to learn about beautiful soup and Web scraping. Not gonna lie. It is quite fascinating. The process of actually going about and looking at websites and extracting. The right information is a process that is intriguing to me. I feel that if I'm able toUtilize all the skills that I've learned thus far and actually document them properly so I can easily access them on a later point that would be awesome.