Day 46
Today was honestly shit. I really tried to understand how to use the Python library Spotipy. I seriously tried my hardest however, I just could not get it to function properly. It is so obtuse, so stupidly, made and so difficult to work with that. I honestly give up. If I have to make a Spotify playlist creator, I will just use the Spotify web API. And not use this honestly terrible library.