Day 47 and 48.5
It was a bit stupid of me to just forget to upload the work that I did yesterday but it's OK. Working with selenium and it's a bit finicky but not as annoying as the Spotipy library. I'm still going to work with that and finish it but for now I think this was pretty good.