Day 49
I'll be honest some of the projects in this course are a bit lame. The concept of just applying for jobs through LinkedIn is good but it's just really finicky. Going about it like that I just don't find it to be all that fun. And the next day is supposed to be an automated Tinder bot essentially. I'll be honest that's really some thing I don't wanna. I don't wanna set up a tinder account and actually use it. I'll try and figure out to do something else tomorrow or I'll just do day 51 instead of day 50. But I will say the projects are becoming a little bit obtuse and lame. I'll just try and find some other interesting things for me to do with the skills that I've learned.