Day 50
It's surprising, but I was actually able to complete this project. It was a bit difficult to really get going, but I somehow managed to figure it out and I'm not gonna lie. I'm genuinely proud of the result. I do think that if I am able to increase my skills using selenium,I'll have a much better time. But for now I'm happy.