Day 51 and 52
I unfortunately forgot to push my work from yesterday, but I'll do that today instead. So yesterday I mainly just worked on the capstone project of finding the apartments in San Francisco (the top rated ones by the way) and putting them in a spreadsheet. Now I was supposed to put them in a specific Google spreadsheet by entering the information in a Google form but for the life of me I could not get it to work. It was so hard for me to just figure out and I kept failing over and over again so I just decided to convert the data that I got into a CSV file and Collett a day. I do think I will come back to this and properlysubmit it to a Google spreadsheet and not go through the Google form thing but for now I think I did good. Today, however, I got to start working with Flask. Now this is a framework that is similar to react and angular, but it's basically the Python version of them. I do think that I will need to get used to multiple different types of frameworks and that will be good. Let's see how it goes but the starting actually looks pretty good.