Day 53
Finally, I got to actually work with flask. Not gonna lie. It is a very interesting application. It's not as difficult as I initially thought it was. It just works on very simple rules and you have to follow those rules very closely. However, the rest of the actual framework just plays off, simple python rules that I've been learning since the start of the summer.