Day 65
This was the final section about the blog website. Adding the functionality to be able to login and register is the final step and I'm not done yet. However, I will finish it. I also feel like the next few github commits are going to be the last major ones and then they're going to become more sporadic.